refactor: new profile data manager
